HandleRequests/requirements.txt
Django==2.1.1
gunicorn==19.6.0
django-tastypie
djangorestframework
pygments
requests
pandas
sklearn
